+++ micro-points/src/main/java/com/wkclz/micro/points/bean/resp/PointsIssueResp.java
@@ -3,6 +3,8 @@ package com.wkclz.micro.points.bean.resp;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 积分发放出参
  */
@@ -14,12 +16,12 @@ public class PointsIssueResp {
     private String flowNo;
 
     @Schema(description = "本次发放积分数")
-    private Integer points;
+    private BigDecimal points;
 
     @Schema(description = "钱包可用积分余额")
-    private Integer availablePoints;
+    private BigDecimal availablePoints;
 
     @Schema(description = "钱包历史总获得积分")
-    private Integer totalEarnedPoints;
+    private BigDecimal totalEarnedPoints;
 
 }
